+++ common/src/main/resources/db/migration/V4__Add_Visa_Restrictions_Index_Table.sql
@@ -1,6 +1,6 @@
 CREATE TABLE visa_restrictions_index (
   country_code VARCHAR (2) PRIMARY KEY,
-  rank SMALLINT NOT NULL,
-  acc SMALLINT NOT NULL,
-  sharing SMALLINT DEFAULT 0
+  rank INT NOT NULL,
+  acc INT NOT NULL,
+  sharing INT DEFAULT 0
 );
